feat(csharp/src/Drivers/Databricks): Integrate ProxyConfigurations with Drivers (#2794)
Integrates proxy logic with the rest of the Spark-Driver. Also applies
to Cloud-fetch and OAuth2 fetch.

Other drivers, like Impala driver, should be compatible with the change.
However, it is not tested - so I marked in the README that it's not
fully complete

To test, add the following to `DATABRICKS_TEST_CONFIG_FILE` environment
variable:
```
  "http_options": {
    "proxy": {
      "use_proxy": "true",
      "proxy_host": "...",
      "proxy_port":..,
      "proxy_auth": "...", // if using basic auth
      "proxy_uid": "...",
      "proxy_pwd": "..."
    },
    "tls": { // you may need to enable depending on your proxy, or you need to trust the cert
      "enabled": ,
      "allow_self_signed": 
    }
  }
```

And inspect proxy traffic. Verified that it works in driver testing and
Powerbi testing